% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountCompCases.R
\name{CountCompCases}
\alias{CountCompCases}
\alias{print.CountCompCases}
\title{Count Complete Cases

Return for each variable of a data frame the number of missing values and
the complete cases to be expected if this variable would be omitted.}
\usage{
CountCompCases(x)
}
\arguments{
\item{x}{a data.frame containg the data.}

\item{digits}{the number of digits to be used when printing the results.}

\item{\\dots}{the dots are not further used.}
}
\value{
A list with three elements. The first gives the number of rows, the
second the number of complete cases for the whole data frame. The third
element \code{tab} contains the data for the single variables.
}
\description{
Count Complete Cases

Return for each variable of a data frame the number of missing values and
the complete cases to be expected if this variable would be omitted.
}
\examples{
                                                                    

CountCompCases(d.pizza)

                                             
}
\seealso{
\code{\link{PlotMiss}}, \code{\link{CompleteColumns}},
\code{\link{complete.cases}}, \code{\link{is.na}}, \code{\link{na.omit}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
