% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_CstatCI.R
\name{CstatCI}
\alias{CstatCI}
\title{Confidence Intervals for the C Statistic (AUC)}
\usage{
CstatCI(
  object,
  conf.level = 0.95,
  sides = c("two.sided", "left", "right"),
  ...
)
}
\arguments{
\item{object}{the model object as returned by glm.}

\item{conf.level}{confidence level of the interval.}

\item{sides}{a character string specifying the side of the confidence
interval, must be one of \code{"two.sided"} (default), \code{"left"} or
\code{"right"}. \code{"left"} would be analogue to a hypothesis of
\code{"greater"} in a \code{t.test}. You can specify just the initial
letter.}

\item{...}{further arguments are passed to the \code{\link{boot}} function.
Supported arguments are \code{type} (\code{"norm"}, \code{"basic"},
\code{"stud"}, \code{"perc"}, \code{"bca"}), \code{parallel} and the number
of bootstrap replicates \code{R}. If not defined those will be set to their
defaults, being \code{"basic"} for \code{type}, option
\code{"boot.parallel"} (and if that is not set, \code{"no"}) for
\code{parallel} and \code{999} for \code{R}.}
}
\value{
a numeric vector with 3 elements: \item{mean}{mean}
\item{lwr.ci}{lower bound of the confidence interval} \item{upr.ci}{upper
bound of the confidence interval}
}
\description{
Calculate bootstrap intervals for the C statistic (Area under the curve AUC),
based on a glm.
}
\examples{

utils::data(Pima.te, package = "MASS")
r.logit <- glm(type ~ ., data=Pima.te, family="binomial")

# calculate Brier score with confidence intervals
Cstat(r.logit)
CstatCI(r.logit, R=99)  # use higher R in real life!
}
\seealso{
\code{\link{BrierScore}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{confidence}
\keyword{interval}
