\name{day.name}
\alias{day.name}
\alias{day.abb}
\docType{data}

\title{Build-in Constants Extension
}

\description{There's a small number of built-in constants in \code{R}. We have month.name and month.abb but nothing similar for days. Here it is.
}

\usage{
day.name
day.abb
}

\seealso{\code{\link{month.name}}, \code{\link{month.abb}}
%% ~~objects to See Also as , ~~~
}

\keyword{datasets}
