% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downloads.R
\name{Downloads}
\alias{Downloads}
\title{Return Downloaded File in XL- or CSV Format}
\usage{
Downloads(fname, ...)
}
\arguments{
\item{fname}{the filename, if no extension specified the first file found
will be used.}

\item{\dots}{are passed on to the open functions
\code{\link[readxl]{readxl}}() or \code{\link[readr]{read.table()}}.}
}
\value{
the file
}
\description{
The function searches for a file in the downloads folder and returns the
file, provided it is in a readable Excel- or data format.
}
\examples{
\dontrun{
Downloads("yourfilename")
Downloads("yourfilename.xls")
Downloads("yourfilename.txt", sep=",", header=TRUE)
}

}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
