% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mbeta}
\alias{mbeta}
\title{Mean and Variance of the Beta Distribution}
\usage{
mbeta(shape1, shape2)
}
\arguments{
\item{shape1}{Alpha parameter}

\item{shape2}{Beta parameter}
}
\value{
List with mean and variance
}
\description{
Formula:
\eqn{\mu = \frac{\alpha}{\alpha + \beta}}
\eqn{\mathrm{Var}(X) = \frac{\alpha \cdot \beta}{(\alpha + \beta)^2 (\alpha + \beta + 1)}}
}
\examples{
mbeta(shape1 = 2, shape2 = 3)
}
\seealso{
\code{\link[stats]{dbeta}}
}
