% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mbinom}
\alias{mbinom}
\title{Mean and Variance of the Binomial Distribution}
\usage{
mbinom(size, prob)
}
\arguments{
\item{size}{Number of trials}

\item{prob}{Probability of success}
}
\value{
List with mean and variance
}
\description{
Formula:
\eqn{\mu = n \cdot p}
\eqn{\mathrm{Var}(X) = n \cdot p \cdot (1 - p)}
}
\examples{
mbinom(size = 10, prob = 0.3)
}
\references{
Casella, G. and Berger, R.L. (2002). Statistical Inference. Duxbury.
}
\seealso{
\code{\link[stats]{dbinom}}
}
