% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mf}
\alias{mf}
\title{Mean and Variance of the F Distribution}
\usage{
mf(df1, df2)
}
\arguments{
\item{df1}{Numerator degrees of freedom}

\item{df2}{Denominator degrees of freedom}
}
\value{
List with mean and variance
}
\description{
Formula:
\eqn{\mu = \frac{df[2]}{df[2] - 2} \quad (df[2] > 2)}
\eqn{\mathrm{Var}(X) = \frac{2 df[2]^2 (df[1] + df[2] - 2)}{df[1] (df[2] - 2)^2 (df[2] - 4)} \quad (df[2] > 4)}
}
\examples{
mf(df1 = 5, df2 = 10)
}
\seealso{
\code{\link[stats]{df}}
}
