% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mgamma}
\alias{mgamma}
\title{Mean and Variance of the Gamma Distribution}
\usage{
mgamma(shape, rate)
}
\arguments{
\item{shape}{Shape parameter}

\item{rate}{Rate parameter}
}
\value{
List with mean and variance
}
\description{
Formula:
\eqn{\mu = \frac{\mathrm{shape}}{\mathrm{rate}}}
\eqn{\mathrm{Var}(X) = \frac{\mathrm{shape}}{\mathrm{rate}^2}}
}
\examples{
mgamma(shape = 2, rate = 0.5)
}
\references{
Evans, M., Hastings, N., and Peacock, B. (2000). Statistical Distributions. Wiley.
}
\seealso{
\code{\link[stats]{dgamma}}
}
