% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mgeom}
\alias{mgeom}
\title{Mean and Variance of the Geometric Distribution}
\usage{
mgeom(prob)
}
\arguments{
\item{prob}{Probability of success}
}
\value{
List with mean and variance
}
\description{
Formula:
\eqn{\mu = \frac{1 - p}{p}}
\eqn{\mathrm{Var}(X) = \frac{1 - p}{p^2}}
}
\examples{
mgeom(prob = 0.2)
}
\seealso{
\code{\link[stats]{dgeom}}
}
