% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mlnorm}
\alias{mlnorm}
\title{Mean and Variance of the Log-Normal Distribution}
\usage{
mlnorm(meanlog, sdlog)
}
\arguments{
\item{meanlog}{Mean on the log scale}

\item{sdlog}{Standard deviation on the log scale}
}
\value{
List with mean and variance
}
\description{
Formula:
\eqn{\mu = \exp(\mu_{log} + 0.5 \cdot \sigma_{log}^2)}
\eqn{\mathrm{Var}(X) = (\exp(\sigma_{log}^2) - 1) \cdot \exp(2\mu_{log} + \sigma_{log}^2)}
}
\examples{
mlnorm(meanlog = 0, sdlog = 1)
}
\seealso{
\code{\link[stats]{dlnorm}}
}
