% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mnbinom}
\alias{mnbinom}
\title{Mean and Variance of the Negative Binomial Distribution}
\usage{
mnbinom(size, prob)
}
\arguments{
\item{size}{Number of successes}

\item{prob}{Probability of success}
}
\value{
List with mean and variance
}
\description{
Formula:
\eqn{\mu = r \cdot \frac{1 - p}{p}}
\eqn{\mathrm{Var}(X) = r \cdot \frac{1 - p}{p^2}}
}
\examples{
mnbinom(size = 5, prob = 0.4)
}
\references{
Casella, G. and Berger, R.L. (2002). Statistical Inference. Duxbury.
}
\seealso{
\code{\link[stats]{dnbinom}}
}
