% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mnorm}
\alias{mnorm}
\title{Mean and Variance of the Normal Distribution}
\usage{
mnorm(mean, sd)
}
\arguments{
\item{mean}{Mean of the distribution}

\item{sd}{Standard deviation}
}
\value{
List with mean and variance
}
\description{
Formula:
\eqn{\mu = \mu}
\eqn{\mathrm{Var}(X) = \sigma^2}
}
\examples{
mnorm(mean = 0, sd = 1)
}
\seealso{
\code{\link[stats]{dnorm}}
}
