% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mt}
\alias{mt}
\title{Mean and Variance of the t-Distribution}
\usage{
mt(df)
}
\arguments{
\item{df}{Degrees of freedom}
}
\value{
List with mean and variance
}
\description{
Formula:
\eqn{\mu = 0 \quad (df > 1)}
\eqn{\mathrm{Var}(X) = \frac{df}{df - 2} \quad (df > 2)}
}
\examples{
mt(df = 5)
}
\seealso{
\code{\link[stats]{dt}}
}
