% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringsAsFactors.R
\name{StringsAsFactors}
\alias{StringsAsFactors}
\title{Convert Character Columns to Factors}
\usage{
StringsAsFactors(x, columns = NULL)
}
\arguments{
\item{x}{the data.frame}

\item{columns}{names or indexes of the columns to be converted;
negative values can be used to omit columns.}
}
\value{
the given data.frame including the converted factors
}
\description{
Strings as factors have recently been downgraded in base R's
\code{\link{data.frame}()} function. However,
it is still usually a good idea to encode string variables
as factors. This function helps to convert some or
all columns of a data.frame to factors.
}
\examples{
# get some data
d.dat <- data.frame(char_x=LETTERS[1:5],
                    char_y=LETTERS[6:10],
                    n=1:5)

# all characters
str(StringsAsFactors(d.dat))
# only char_y
str(StringsAsFactors(d.dat, columns="char_y"))
# only char_x
str(StringsAsFactors(d.dat, columns="char_x"))

# all characters, besides 2 (second column, so "char_y")
str(StringsAsFactors(d.dat, columns=-2))

}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
