\name{day.name}
\alias{day.name}
\alias{day.abb}
\docType{data}

\title{Build-in Constants Extension
}

\description{There's a small number of built-in constants in \code{R}. We have \code{\link{month.name}} and \code{\link{month.abb}} but nothing similar for weekdays. Here it is.
}

\usage{
day.name
day.abb
}

\details{
The following constants are available in DescTools:
\itemize{
    \item \code{day.name}: the English names for the day of the week (Monday, Tuesday,   Wednesday, Thursday,  Friday,   Saturday, Sunday);
    \item \code{day.abb}: the three-letter abbreviations for the
    English day names (Mon, Tue, Wed, Thu, Fri, Sat, Sun);
  }

}

\seealso{\code{\link{month.name}}, \code{\link{month.abb}}
%% ~~objects to See Also as , ~~~
}

\keyword{datasets}
