% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.r
\name{Mean}
\alias{Mean}
\alias{Mean.default}
\alias{Mean.Freq}
\title{(Weighted) Arithmetic Mean}
\usage{
Mean(x, ...)

\method{Mean}{Freq}(x, breaks, ...)

\method{Mean}{default}(x, weights = NULL, trim = 0, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An object.  Currently there are methods for numeric/logical vectors
and \link[=Dates]{date}, \link{date-time} and \link{time interval} objects.
Complex vectors are allowed for \code{trim = 0} only.}

\item{\dots}{further arguments passed to or from other methods.}

\item{breaks}{breaks for calculating the mean for classified data as
composed by \code{\link{Freq}}.}

\item{weights}{non-negative numeric vector of weights of the same length as \code{x}
interpreted as frequency (replication) weights.
Observations with larger weights contribute
more strongly to the empirical distribution.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each
end of \code{x} before the mean is computed.  Values of trim outside that
range are taken as the nearest endpoint.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\value{
If \code{trim} is zero (the default), the arithmetic mean of the
values in \code{x} is computed, as a numeric or complex vector of length
one.  If \code{x} is not logical (coerced to numeric), numeric (including
integer) or complex, \code{NA_real_} is returned, with a warning.

If \code{trim} is non-zero, a symmetrically trimmed mean is computed with a
fraction of \code{trim} observations deleted from each end before the mean
is computed.
}
\description{
Generic function for the (trimmed) arithmetic mean, possibly with given
weights.
}
\details{
The argument \code{weights} is interpreted as frequency (replication)
weights. Conceptually, this corresponds to computing the statistic
on a reweighted empirical distribution, where observations with larger
weights represent a higher frequency in the population.

\bold{Note:}\verb{    }
Analytic (precision) weights, which assume observation-specific error
variances or likelihood-based weighting, are deliberately not supported.
This design ensures that all weighted statistics remain well-defined
for ordinal, robust, and distribution-based measures such as medians,
quantiles, and measures of dispersion.

\code{trim} and \code{weights} can't be used together at the same time.
}
\examples{

x <- c(0:10, 50)
xm <- Mean(x)
c(xm, Mean(x, trim = 0.10))

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) \emph{The
New S Language}.  Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link{weighted.mean}}, \code{\link{mean.POSIXct}},
\code{\link{colMeans}} for row and column means.
}
\keyword{univar}
