% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{mt}
\alias{mt}
\title{Mean and Variance of the t-Distribution}
\usage{
mt(df)
}
\arguments{
\item{df}{degrees of freedom (\eqn{> 0}, maybe non-integer). \code{df = Inf} is allowed.}
}
\value{
List with mean and variance
}
\description{
The \eqn{t} distribution with \code{df} \eqn{= \nu}{= n} degrees of
freedom has mean and variance of:
\eqn{\mu = 0 \quad (\nu > 1)}
\eqn{\mathrm{Var}(X) = \frac{\nu}{\nu - 2} \quad (\nu > 2)}
}
\examples{
mt(df = 5)
}
\references{
Casella, G. and Berger, R.L. (2002). Statistical Inference. Duxbury.
}
\seealso{
\code{\link[stats]{dt}}
}
