% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringsAsFactors.R
\name{StringsAsFactors}
\alias{StringsAsFactors}
\title{Convert Character Columns to Factors}
\usage{
StringsAsFactors(x, columns = NULL)
}
\arguments{
\item{x}{the data.frame}

\item{columns}{names or indexes of the columns to be converted;
negative values can be used to omit columns.}
}
\value{
the given data.frame including the converted factors
}
\description{
A helper function to convert some or all columns of a data.frame to factors.
}
\examples{
# get some data
d.dat <- data.frame(char_x=LETTERS[1:5],
                    char_y=LETTERS[6:10],
                    n=1:5)

# all characters
StringsAsFactors(d.dat)
# only char_y
StringsAsFactors(d.dat, columns="char_y")
# only char_x
StringsAsFactors(d.dat, columns="char_x")

# all characters, besides 2 (second column, so "char_y")
StringsAsFactors(d.dat, columns=-2)

}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
