\name{SysInfo}
\alias{SysInfo}
\alias{FindRProfile}

\title{System Information}
\description{SysInfo is a convenience function to compile some information about the
computing system and environment used.}
\usage{SysInfo()
FindRProfile()
}

\details{The function \code{SysInfo} is mainly used to save the system environment information
in ncdf files containing the results of some calculations. \cr\cr
\code{FindRProfile} returns path candidates where the profile could be found.
}
\value{character string with all version and system information of the current R system}

\author{Jannis v. Buttlar <jbuttlar@bgc-jena.mpg.de>, Andri Signorell <andri@signorell.net>}





